agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-core>=1.0.0
openai>=1.50.0
httpx>=0.27.0
pydantic>=2.8.0
python-dotenv>=1.0.0
